+++ src/Core/Checkout/Cart/Collector/AddTaxCollector.php
@@ -2,21 +2,35 @@
 
 declare(strict_types=1);
 
-namespace solu1TaxJar\Core\Checkout\Cart\Collector;
+namespace ITGCoTax\Core\Checkout\Cart\Collector;
 
+use AllowDynamicProperties;
+use ITGCoTax\Core\Checkout\Cart\Extension\TaxExtensionStruct;
 use Psr\Cache\CacheItemPoolInterface;
 use Shopware\Core\Checkout\Cart\Cart;
 use Shopware\Core\Checkout\Cart\CartBehavior;
 use Shopware\Core\Checkout\Cart\CartProcessorInterface;
 use Shopware\Core\Checkout\Cart\LineItem\CartDataCollection;
 use Shopware\Core\Checkout\Cart\LineItem\LineItem;
+use Shopware\Core\Checkout\Cart\Price\QuantityPriceCalculator;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
 
+#[AllowDynamicProperties]
 class AddTaxCollector implements CartProcessorInterface
 {
+    /**
+     * @var EntityRepository
+     */
+
+
+    /**
+     * @var QuantityPriceCalculator
+     */
+
     /**
      * @var EntityRepository
      */
@@ -74,16 +88,18 @@ class AddTaxCollector implements CartProcessorInterface
     private function getTaxProviderClass($taxRuleId, SalesChannelContext $context)
     {
         if ($taxRuleId) {
-            $criteria = new Criteria([$taxRuleId]);
+            $criteria = new Criteria();
+            $criteria->addFilter(new EqualsAnyFilter('id', [0 => $taxRuleId]));
             $taxRules = $this->taxRepository->search($criteria, $context->getContext());
             foreach ($taxRules as $taxRule) {
-                if ($taxRule->getExtension('taxExtension') &&
-                    $taxProviderId = $taxRule->getExtension('taxExtension')->getProviderId()) {
-                    $criteria = new Criteria([$taxProviderId]);
+                /** @var TaxExtensionStruct|null $extension */
+                $extension = $taxRule->getExtension('taxExtension');
+                if ($extension && $extension->getProviderId()) {
+                    $criteria = new Criteria([$extension->getProviderId()]);
                     $taxProviders = $this->taxProviderRepository->search($criteria, $context->getContext());
                     foreach ($taxProviders as $taxProvider) {
-                        if ($taxProvider->getBaseClass()) {
-                            $taxCalculatorClass = $taxProvider->getBaseClass();
+                        $taxCalculatorClass = $taxProvider->get('baseClass');
+                        if ($taxCalculatorClass) {
                             return new $taxCalculatorClass(
                                 $this->systemConfigService,
                                 $this->taxJarLogRepository,