+++ src/Core/Content/TaxLog/TaxLogDefinition.php
@@ -1,7 +1,12 @@
 <?php
+/**
+ * Copyright ©2021 ITG Commerce Ltd., Inc. All rights reserved.
+ * See COPYING.txt for license details.
+
+ */
 declare(strict_types=1);
 
-namespace solu1TaxJar\Core\Content\TaxLog;
+namespace ITGCoTax\Core\Content\TaxLog;
 
 use Shopware\Core\Framework\DataAbstractionLayer\EntityDefinition;
 use Shopware\Core\Framework\DataAbstractionLayer\Field\CreatedAtField;
@@ -16,7 +21,7 @@ use Shopware\Core\Framework\DataAbstractionLayer\FieldCollection;
 
 class TaxLogDefinition extends EntityDefinition
 {
-    public const ENTITY_NAME = 's25_taxjar_log';
+    public const ENTITY_NAME = 'itg_taxjar_log';
 
     /**
      * @return string