+++ src/Migration/Migration1650801564MyQuery.php
@@ -1,8 +1,12 @@
 <?php
+/**
+ * Copyright ©2021 ITG Commerce Ltd., Inc. All rights reserved.
+ * See COPYING.txt for license details.
 
+ */
 declare(strict_types=1);
 
-namespace solu1TaxJar\Migration;
+namespace ITGCoTax\Migration;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
@@ -20,7 +24,7 @@ class Migration1650801564MyQuery extends MigrationStep
     {
         $query = /** @lang SQL */
             <<<SQL
-          CREATE TABLE IF NOT EXISTS `s25_tax_service_provider` (
+          CREATE TABLE IF NOT EXISTS `itg_tax_service_provider` (
           `id` binary(16) NOT NULL,
           `name` varchar(512) DEFAULT NULL,
           `base_class` text,
@@ -31,7 +35,7 @@ class Migration1650801564MyQuery extends MigrationStep
        ;
 SQL;
         $connection->executeStatement($query);
-        $connection->executeStatement('TRUNCATE TABLE `s25_tax_service_provider`');
+        $connection->executeStatement('TRUNCATE TABLE `itg_tax_service_provider`');
         $this->addTaxjarServiceProvider($connection);
     }
 
@@ -45,7 +49,7 @@ SQL;
         $taxJarDataEntry = $this->getTaxJarData();
         $taxJarDataEntry['id'] = Uuid::randomBytes();
         $taxJarDataEntry['created_at'] = date('Y:m:d H:i:s', time());
-        $connection->insert('s25_tax_service_provider', $taxJarDataEntry);
+        $connection->insert('itg_tax_service_provider', $taxJarDataEntry);
     }
 
     /**
@@ -55,12 +59,12 @@ SQL;
     {
         return [
             'name' => 'TaxJar',
-            'base_class' => '\solu1TaxJar\Core\TaxJar\Calculator',
+            'base_class' => '\ITGCoTax\Core\TaxJar\Calculator',
         ];
     }
 
     public function updateDestructive(Connection $connection): void
     {
-        $connection->executeStatement('DROP TABLE IF EXISTS `s25_tax_service_provider`');
+        $connection->executeStatement('DROP TABLE IF EXISTS `itg_tax_service_provider`');
     }
 }