+++ src/Migration/Migration1651411372MyQuery.php
@@ -1,7 +1,12 @@
 <?php
+/**
+ * Copyright ©2021 ITG Commerce Ltd., Inc. All rights reserved.
+ * See COPYING.txt for license details.
+
+ */
 declare(strict_types=1);
 
-namespace solu1TaxJar\Migration;
+namespace ITGCoTax\Migration;
 
 use Doctrine\DBAL\Connection;
 use Shopware\Core\Framework\Migration\MigrationStep;
@@ -17,7 +22,7 @@ class Migration1651411372MyQuery extends MigrationStep
     {
        $query = /** @lang SQL */
             <<<SQL
-          CREATE TABLE IF NOT EXISTS `s25_tax_provider` (
+          CREATE TABLE IF NOT EXISTS `itg_tax_provider` (
           `id` binary(16) NOT NULL,
           `tax_id` binary(16) NOT NULL,
           `provider_id` binary(16) NOT NULL,
@@ -33,6 +38,6 @@ SQL;
 
     public function updateDestructive(Connection $connection): void
     {
-        $connection->executeStatement('DROP TABLE IF EXISTS `s25_tax_provider`');
+        $connection->executeStatement('DROP TABLE IF EXISTS `itg_tax_provider`');
     }
 }