+++ src/Storefront/Controller/ExportLogController.php
@@ -1,9 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace solu1TaxJar\Storefront\Controller;
+namespace ITGCoTax\Storefront\Controller;
 
-use Exception;
-use solu1TaxJar\Core\Content\TaxLog\TaxLogEntity;
+
+use ITGCoTax\Core\Content\TaxLog\TaxLogEntity;
 use Shopware\Core\Framework\DataAbstractionLayer\Dbal\Common\RepositoryIterator;
 use Symfony\Component\HttpFoundation\Request;
 use Shopware\Core\Framework\Context;
@@ -17,10 +17,10 @@ class ExportLogController
     /**
      * @var EntityRepository
      */
-    private $taxJarLogRepository;
+    private EntityRepository $taxJarLogRepository;
 
     /**
-     * @param EntityRepository $taxJarLogRepository
+     * @var EntityRepository $taxJarLogRepository
      */
     public function __construct(
         EntityRepository $taxJarLogRepository
@@ -44,11 +44,11 @@ class ExportLogController
                 $response[] = [
                     'customerName' => $taxJarLog->getCustomerName(),
                     'customerEmail' => $taxJarLog->getCustomerEmail(),
-                    'orderNumber' => $taxJarLog->getOrderNumber()??'',
-                    'orderId' => $taxJarLog->getOrderId()??'',
-                    'remoteIp' => $taxJarLog->getRemoteIp()?? '',
+                    'orderNumber' => $taxJarLog->getOrderNumber(),
+                    'orderId' => $taxJarLog->getOrderId(),
+                    'remoteIp' => $taxJarLog->getRemoteIp(),
                     'request' => str_replace('"', "'", $taxJarLog->getRequest()),
-                    'response' => str_replace('"', "'", $taxJarLog->getResponse()),
+                    'response' => str_replace('"', "'",$taxJarLog->getResponse()),
                     'createdAt' => $taxJarLog->getCreatedAt()->format(DATE_RFC3339_EXTENDED)
                 ];
             }