+++ Changelog.md
@@ -2,8 +2,17 @@
 
 All notable changes to this tax-jar-shopware-6 plugin will be documented in this file.
 
-## [1.0.0] - 2025-09-24
+## [1.0.3] - 2025-02-04
 
-###
+### Added
+- Added Order Number For Order Cancellation And Order Refund Request To Tax Jar
+- Added Order ID In Tax Jar Log List 
+- Added Order ID In Filter Section Of Tax Jar Log
+- Added Order ID And Order Number in Tax Jar Export
 
-###
+### Fixed
+
+- Fixed Missing Refresh Icon On Tax Jar Log Page.
+- Fixed Missing Product Tax Code In Order Amendment Request
+- Column not found: 1054 Unknown column 'order_id' while taxjar request logging
+